메일 Blob 테이블에만 있는 레코드 삭제 시 mailboxId와 MailUid가 null인 경우에 대한 예외처리 추가
  - 원인을 정확히 알 수 없으나 료비에서 mailboxId와 mailUid가 null인 레코드가 발생하여 추가함
  - 예외처리를 하지 않으면 deleteMailBlob 스케쥴러 처리 전체가 실패하여 레코드 삭제가 되지 않는
    문제가 발생함

+++ src/main/java/egovframework/ezEKP/ezEmail/dao/EzEmailDAO.java
@@ -69,24 +69,31 @@ public class EzEmailDAO extends EgovAbstractDAO {
 		delete("EzEmailDAO.deleteOrphanedMailBlob", mailBlobVO);
 		
 		long sleepTime = 500;
-		long mailboxId = mailBlobVO.getMailBoxId();
-		long mailUid = mailBlobVO.getMailUid();
-		String headerPath = ezEmailUtil.getMailHeaderPath(mailboxId, mailUid);
-		String bodyPath = ezEmailUtil.getMailBodyPath(mailboxId, mailUid);
-		File headerFile = new File(headerPath);
-		File bodyFile = new File(bodyPath);
 		
-		if (headerFile.exists()) {
-			headerFile.delete();
+		if (mailBlobVO.getMailBoxId() != null && mailBlobVO.getMailUid() != null) {
+			long mailboxId = mailBlobVO.getMailBoxId();
+			long mailUid = mailBlobVO.getMailUid();
+			String headerPath = ezEmailUtil.getMailHeaderPath(mailboxId, mailUid);
+			String bodyPath = ezEmailUtil.getMailBodyPath(mailboxId, mailUid);
+			File headerFile = new File(headerPath);
+			File bodyFile = new File(bodyPath);
+			
+			if (headerFile.exists()) {
+				headerFile.delete();
+				
+				sleepTime = 0;
+			}
+			
+			if (bodyFile.exists()) {
+				bodyFile.delete();
+				
+				sleepTime = 0;
+			}		
+		} else {
+			logger.debug("deleteOrphanedMailBlob mailboxId=" + mailBlobVO.getMailBoxId() + ",mailUid=" + mailBlobVO.getMailUid());
 			
 			sleepTime = 0;
 		}
-		
-		if (bodyFile.exists()) {
-			bodyFile.delete();
-			
-			sleepTime = 0;
-		}		
 		
 		return sleepTime;
 	}
